api-auth/requirements.txt
fastapi==0.115.5
nest_asyncio==1.6.0
PyJWT==2.10.1
starlette==0.41.3
uvicorn==0.32.1
python-multipart
cryptography